update refreshToken with redis

+++ src/main/java/com/example/demo_spring_boot/service/impl/RoleServiceImpl.java
@@ -5,7 +5,9 @@ import com.example.demo_spring_boot.model.entity.Role;
 import com.example.demo_spring_boot.repository.RoleRepository;
 import com.example.demo_spring_boot.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -49,6 +51,9 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void deleteRole(Long id) {
+        if (!roleRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User không tồn tại");
+        }
         roleRepository.deleteById(id);
     }
 }
